feat: Persistent storage (#13)

+++ kvnode/src/actor.rs
@@ -1,62 +1,85 @@
-use serde::Deserialize;
-use serde::Serialize;
+use std::path::PathBuf;
+
+use rkyv::Archive;
 use tokio::sync::mpsc;
 
 use tokio::sync::oneshot;
+use tracing::info;
 
+use crate::persistence::WalManager;
 use crate::storage::KVStore;
 use crate::storage::StorageInterface;
 
-#[derive(Serialize, Deserialize)]
+#[derive(Archive, rkyv::Serialize, rkyv::Deserialize, serde::Deserialize, Debug, PartialEq)]
 pub struct PutRequest {
     pub key: String,
     pub value: String,
 }
 
+#[derive(Archive, rkyv::Serialize, rkyv::Deserialize, serde::Deserialize, Debug, PartialEq)]
 pub enum StoreCommand {
-    Get {
-        key: String,
-        reply: oneshot::Sender<Option<String>>,
-    },
-    Put(PutRequest),
-    Delete {
-        key: String,
-    },
+    Get { key: String },
+    Put { record: PutRequest },
+    Delete { key: String },
 }
 
+pub type ActorMessage = (StoreCommand, Option<oneshot::Sender<Option<String>>>);
+
 // Actor controls access to the internal local store by message passing
-pub async fn node_actor(mut rx: mpsc::Receiver<StoreCommand>) {
+pub async fn node_actor(mut rx: mpsc::Receiver<ActorMessage>, wal_path: PathBuf) {
     let mut store = KVStore::default();
-    while let Some(cmd) = rx.recv().await {
-        match cmd {
-            StoreCommand::Get { key, reply } => {
-                let _ = reply.send(store.get(&key));
+    let wal = WalManager::new(wal_path).await;
+    while let Some(msg) = rx.recv().await {
+        info!(message = "Actor received:", message = ?msg);
+
+        match msg {
+            (StoreCommand::Get { key }, reply) if reply.is_some() => {
+                if let Some(tx) = reply {
+                    let _ = tx.send(store.get(&key));
+                }
             }
-            StoreCommand::Put(PutRequest { key, value }) => {
-                store.put(&key, &value);
+            (cmd @ StoreCommand::Put { .. }, None) => {
+                // Log
+                wal.write(&cmd).await;
+                if let StoreCommand::Put {
+                    record: PutRequest { key, value },
+                } = cmd
+                {
+                    store.put(&key, &value);
+                }
             }
-            StoreCommand::Delete { key } => {
-                store.delete(&key);
+            (cmd @ StoreCommand::Delete { .. }, None) => {
+                // Log
+                wal.write(&cmd).await;
+                if let StoreCommand::Delete { key } = cmd {
+                    store.delete(&key);
+                }
+            }
+            (msg, return_channel) => {
+                panic!("illegal {:?} - {:#?}", msg, return_channel);
             }
         }
     }
+
+    // wal.replay().await;
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
 
-    async fn get_key(tx: &mpsc::Sender<StoreCommand>, key: String) -> Option<String> {
+    async fn get_key(tx: &mpsc::Sender<ActorMessage>, key: String) -> Option<String> {
         // Check key not exist
         let (itx, irx) = oneshot::channel();
-        let _ = tx.send(StoreCommand::Get { key, reply: itx }).await;
+        let _ = tx.send((StoreCommand::Get { key }, Some(itx))).await;
         irx.await.unwrap()
     }
 
     #[tokio::test]
     async fn actor_fileops() {
-        let (tx, rx) = mpsc::channel::<StoreCommand>(32);
-        tokio::spawn(node_actor(rx));
+        let (tx, rx) = mpsc::channel::<ActorMessage>(32);
+        let tmp_wal = PathBuf::from("/dev/null");
+        tokio::spawn(node_actor(rx, tmp_wal));
 
         // Check key not exist
         assert!(get_key(&tx, "key".into()).await.is_none());
@@ -64,19 +87,21 @@ mod tests {
         // insert key
         let _ = tx
             .clone()
-            .send(StoreCommand::Put(PutRequest {
-                key: "key".into(),
-                value: "new_value".into(),
-            }))
+            .send((
+                StoreCommand::Put {
+                    record: PutRequest {
+                        key: "key".into(),
+                        value: "new_value".into(),
+                    },
+                },
+                None,
+            ))
             .await;
 
         // Check key exists
         let (itx, irx) = oneshot::channel();
         let _ = tx
-            .send(StoreCommand::Get {
-                key: "key".into(),
-                reply: itx,
-            })
+            .send((StoreCommand::Get { key: "key".into() }, Some(itx)))
             .await;
         let response = irx.await.unwrap();
         assert_eq!(response, Some("new_value".into()));
@@ -85,7 +110,7 @@ mod tests {
         // Remove key
         let _ = tx
             .clone()
-            .send(StoreCommand::Delete { key: "key".into() })
+            .send((StoreCommand::Delete { key: "key".into() }, None))
             .await;
 
         // Check key not exist
